+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Changelog
 
-## ???
+## 1.1.0
 - Support LaTeX compiling using the LaTeX Workshop extension
-- Log to output channel
-- Fix synchronization bug (from project copy to working directory)
+- Improve logs and log to an output channel and a log file
+- Fix various bugs (including a synchronization issue)
+- Some QOL improvements in the configuration file
 
 ## 1.0.1
 Initial release